Also patch `salt/version.py` when making a release.
